Promote activations and gradients to float32 for Hessian outer products
Running EKFAC with a bf16 model hit two distinct failure modes:

1. `torch.einsum` in ShardedMul raised
   `RuntimeError: expected scalar type Float but found BFloat16`
   when the stored eigenvectors (fp32) were combined with the model's bf16
   activations/gradients. Fixed by `.float()`-ing the vector input and
   allocating the result tensors as fp32.

2. The bf16 `a.mT @ a` and `g.mT @ g` outer products accumulated into fp32
   covariance buffers with visible numerical drift. Casting the selected
   activations/gradients to fp32 before the matmul restores the behaviour
   of the earlier non-upstream implementation.

Lifted from the earlier prototype commit 35189d6 ("Fix mixed-precision
handling and add jaxtyping annotations") and pared down to the two files
that still exist in the current bergson layout.

+++ bergson/hessians/kfac.py
@@ -48,7 +48,8 @@ class CovarianceCollector(HookCollectorBase):
         assert mask is not None, "Valid mask not set for forward hook."
 
         # a: [N, S, I], valid_masks: [N, S] -> select valid positions
-        a_bi = a[mask]  # [num_valid, I]
+        # Promote to float32 to avoid bf16 numerical drift in the outer-product accumulator.
+        a_bi = a[mask].float()  # [num_valid, I]
 
         # Compute local covariance
         local_update_ii = a_bi.mT @ a_bi
@@ -72,7 +73,8 @@ class CovarianceCollector(HookCollectorBase):
         mask = self._current_valid_mask
 
         # g: [N, S, O], mask: [N, S] -> select valid positions
-        g_bo = g[mask]  # [num_valid, O]
+        # Promote to float32 to avoid bf16 numerical drift in the outer-product accumulator.
+        g_bo = g[mask].float()  # [num_valid, O]
 
         # Compute local covariance
         local_update_oo = g_bo.mT @ g_bo
